Added 'main' routine to script. Added CLI options.
The script now how a separate 'main' routine, so that it can also be
used as a module.

The correctness of the command line call is checked; there is a proper
error message in the case of wrong usage.

Added an option to change the typescripts directory. The directory is
now created if it doesn't exist already.